# src/main/resources/openapi/api.yaml
openapi: 3.0.1
info:
  title: Student Tutor Booking Management System
  description: This is a system allowing students and tutors to book in-person sessions
    at designated venues on campus.
  termsOfService: http://swagger.io/terms/
  contact:
    email: 222641495@mycput.ac.za
  version: 1.0.0
servers:
  - url: https://localhost:8080
...
paths:
...
  /location:
    get:
      tags:
        - location
      summary: Get all location
      description: Returns all locations in the system
      operationId: getAllLocations
      responses:
        '200':
          description: Retrieved locations successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LocationDTO'
        '400':
          description: Error retrieving locations
          content: { }
    post:
      tags:
        - location
      summary: Add a new venue
      description: This can only be done by the admin.
      operationId: addLocation
      requestBody:
        description: Location object that needs to be added to the system
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationDTO'
        required: true
      responses:
        "201":
          description: Location added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationDTO'
        "400":
          description: Invalid input
          content: {}
      x-codegen-request-body-name: body
  /location/{locationId}:
    get:
      tags:
        - location
      summary: Get a venue by id
      description: This retrieves a location by its id
      operationId: getLocation
      parameters:
        - name: locationId
          in: path
          description: Id of location to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Location found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationDTO'
    put:
      tags:
        - location
      summary: Update an existing location
      description: This can only be done by the admin.
      operationId: updateLocation
      parameters:
        - name: locationId
          in: path
          description: ID of location that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated location object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationDTO'
        required: true
      responses:
        "200":
          description: Location updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationDTO'
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: Location not found
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
        - location
      summary: Delete location
      description: This can only be done by the admin.
      operationId: deleteLocation
      parameters:
        - name: locationId
          in: path
          description: ID of Location that needs to be deleted
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Location deleted successfully
          content: {}
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: Location not found
          content: {}
  /booking:
    get:
      tags:
        - booking
      summary: Get all bookings
      description: Returns all bookings in the system
      operationId: getAllBookings
      responses:
        '200':
          description: Retrieved Bookings successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingDTO'
        '400':
          description: Error retrieving bookings
          content: { }
    post:
      tags:
        - booking
      summary: Create a new booking
      description: This can only be done by the student.
      operationId: createBooking
      requestBody:
        description: BookingDTO object that needs to be added to the system
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingDTO'
        required: true
      responses:
        "201":
          description: BookingDTO created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDTO'
        "400":
          description: Invalid input
          content: {}
      x-codegen-request-body-name: body
  /booking/{bookingId}:
    get:
      tags:
        - booking
      summary: Get booking by ID
      description: Returns a single booking
      operationId: getBookingById
      parameters:
        - name: bookingId
          in: path
          description: ID of booking to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDTO'
        "400":
          description: Invalid ID supplied
          content: {}
        "404":
          description: BookingDTO not found
          content: {}
    put:
      tags:
        - booking
      summary: Update a new booking
      description: This can only be done by the student.
      operationId: updateBooking
      parameters:
        - name: bookingId
          in: path
          description: ID of booking to return
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: BookingDTO object that needs to be updated on the system
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingDTO'
        required: true
      responses:
        "201":
          description: BookingDTO created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDTO'
        "400":
          description: Invalid input
          content: { }
      x-codegen-request-body-name: body
    delete:
      tags:
        - booking
      summary: Delete Booking by ID
      description: Deletes a booking
      operationId: deleteBooking
      parameters:
        - name: bookingId
          in: path
          description: ID of booking to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content: { }
        "400":
          description: Invalid ID supplied
          content: { }
        "404":
          description: BookingDTO not found
          content: { }
  /review:
    get:
      tags:
        - review
      summary: Get all reviews
      description: Returns all admins in the system
      operationId: getAllReviews
      responses:
        '200':
          description: Retrieved reviews successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewDTO'
        '400':
          description: Error retrieving admins
          content: { }
    post:
      tags:
        - review
      summary: Add review
      description: This can only be done by the student.
      operationId: addReview
      requestBody:
        description: ReviewDTO object that needs to be added to the system
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewDTO'
        required: true
      responses:
        "201":
          description: ReviewDTO added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDTO'
        "400":
          description: Invalid input
          content: {}
      x-codegen-request-body-name: body
  /review/{reviewId}:
    get:
      tags:
        - review
      summary: Get review by Id
      description: Fetches data about a particular review
      operationId: getReview
      parameters:
        - name: reviewId
          in: path
          description: Id of review to fetch
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: ReviewDTO retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDTO'
        "404":
          description: Id not found
          content: { }
      x-codegen-request-body-name: body
    put:
      tags:
        - review
      summary: Update existing review
      description: Updates based on ID provided in path
      operationId: updateReview
      parameters:
        - name: reviewId
          in: path
          description: ID of review to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewDTO'
        required: true
      responses:
        "200":
          description: Review updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewDTO'
        "400":
          description: Invalid data
          content: { }
        "404":
          description: Id not found
          content: { }
      x-codegen-request-body-name: body
    delete:
      tags:
        - review
      summary: Delete existing review
      description: deletes based on Id provided
      operationId: deleteReview
      parameters:
        - name: reviewId
          in: path
          description: id of review to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: review deleted successfully
          content: { }
        404:
          description: review not found
          content: { }
components:
  schemas:
...
    BookingDTO:
      required:
        - date
        - locationId
        - studentId
        - subjectId
        - startTime
        - endTime
        - topic
        - tutorId
      type: object
      properties:
        bookingId:
          type: integer
          format: int64
        tutorId:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        subjectId:
          type: integer
          format: int64
        date:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
        locationId:
          type: integer
          format: int64
        topic:
          type: string
    LocationDTO:
      required:
        - building
        - room
      type: object
      properties:
        locationId:
          type: integer
          format: int64
        room:
          type: string
        building:
          type: string
    ReviewDTO:
      required:
        - rating
        - description
        - tutorId
        - studentId
      type: object
      properties:
        reviewId:
          type: integer
          format: int64
        tutorId:
          type: integer
          format: int64
        rating:
          type: integer
          format: int32
        description:
          type: string
        studentId:
          type: integer
          format: int64
...